Update script.js
All Criterias are completed and all functions work smoothly ,but the presentation of each rows is not at the same position. The reason I guessed is that first four contexts "9AM, 10AM, 11AM, 12AM" are two decimals. The "1PM"-"5PM" has only one number. So I will try add a space in the text.